Int32 rename
